feat: flush when shared buffer reaches threshold and refine code

+++ src/storage/src/hummock/mod.rs
@@ -379,7 +379,7 @@ impl StateStore for HummockStorage {
                 })
                 .collect_vec();
 
-            let batch_size = self.shared_buffer_manager.write_batch(batch, epoch).await?;
+            let batch_size = self.shared_buffer_manager.write_batch(batch, epoch).await? as u64;
 
             if !self.options.async_checkpoint_enabled {
                 self.shared_buffer_manager.sync(Some(epoch)).await?;
